Converted DOCTYPE from XHTML to HTML (per "Living Standard")
I'm sad that XHTML is basically dead, but it is.

I've changed the DOCTYPE and added a <meta charset> element.

I've kept the slash/solidus/virgule on all the empty (void) tags,
though. I think it makes more sense. (It should have stayed in as a
requirement, as in XHTML, but backwards compatibility means that is has
to be optional.)

HTML has become the X86 instruction set - full of legacy crap that can
never be taken out.